feat: creación de paquete para base de datos
